Task creation works now
I think I ran into the most insane inconsistencies I've ever seen while fixing this issue.

Mongoose is supposed to automatically generate a value for _id, but for some reason it wasn't doing it, so I had to manually set it, which I DO NOT have to do for creating Users??

Also I stripped the Task schema down to 2 fields so I could fix this, and after I fixed it and reintroduced the other fields, THEN it complained that mongoose wasn't imported in task.js. Even though it was just working without it being imported!