This should fix all our warnings about "keys" and "class" vs "className" stuff
Basically I changed a bunch of lines that said
<div className="overlay" class="fixed pin z-50 overflow-auto flex">

to just be
<div className="fixed pin z-50 overflow-auto flex">

and when I tested the app, everything works the same (background is not clickable when an overlay is shown)
=========================
The other thing I did was make it so we follow conventions for things that use props by giving every element a "key" property. Again, I tested the app and everything works the same.